requirements.txt
instaloader==4.10.3
moviepy==1.0.3
openai-whisper==20231117
SpeechRecognition==3.10.1
